Refactor how core JS modules are organized and make them self referencing (don't use import @vrembem/core)
