+++ packages/core/index.js
@@ -1,24 +1,8 @@
-export * from "./src/js/utilities/localStore";
-export * from "./src/js/utilities/maybeRunMethod";
-export * from "./src/js/utilities/toCamel";
-export * from "./src/js/utilities/toKebab";
-export * from "./src/js/utilities/toMilliseconds";
-export * from "./src/js/utilities/transition";
+export * from "./src/js/utilities";
+export * from "./src/js/helpers";
 
-export * from "./src/js/helpers/cssVar";
-export * from "./src/js/helpers/getConfig";
-export * from "./src/js/helpers/getCustomProps";
-export * from "./src/js/helpers/getElement";
-export * from "./src/js/helpers/getPrefix";
-export * from "./src/js/helpers/getSetting";
-export * from "./src/js/helpers/pluginsArray";
-export * from "./src/js/helpers/setGlobalState";
-
-export * from "./src/js/Breakpoint";
 export * from "./src/js/Collection";
 export * from "./src/js/CollectionEntry";
+export * from "./src/js/Breakpoint";
 export * from "./src/js/FocusTrap";
 export * from "./src/js/themeStore";
-
-export * from "./src/js/plugins/debug";
-export * from "./src/js/plugins/teleport";
