+++ packages/core/plugins.js
@@ -1,0 +1,1 @@
+export * from "./src/js/plugins";
