+++ packages/core/src/js/plugins/index.js
@@ -1,0 +1,2 @@
+export * from "./debug";
+export * from "./teleport";
