+++ packages/core/src/js/themeStore.js
@@ -1,4 +1,5 @@
-import { cssVar, localStore } from "@vrembem/core";
+import { localStore } from "./utilities";
+import { cssVar } from "./helpers";
 
 export function themeStore(options) {
   // Setup the default settings object.
