+++ packages/core/src/js/utilities/getConfig.js
