+++ packages/core/src/js/utilities/index.js
@@ -1,0 +1,9 @@
+export * from "./getConfig";
+export * from "./getElement";
+export * from "./localStore";
+export * from "./maybeRunMethod";
+export * from "./teleport";
+export * from "./toCamel";
+export * from "./toKebab";
+export * from "./toMilliseconds";
+export * from "./transition";
