+++ packages/drawer/index.html
@@ -10,7 +10,7 @@
   <script type="module">
     import "vrembem/dist/index.css";
     import Drawer from "./index.js";
-    import { teleport } from "@vrembem/core";
+    import { teleport } from "@vrembem/core/plugins";
 
     const drawer = new Drawer({
       plugins: [
