+++ packages/modal/index.html
@@ -10,7 +10,7 @@
   <script type="module">
     import "vrembem/dist/index.css";
     import Modal from "./index.js";
-    import { teleport } from "@vrembem/core";
+    import { teleport } from "@vrembem/core/plugins";
 
     const modal = new Modal({
       plugins: [
