+++ packages/vrembem/index.html
@@ -15,6 +15,7 @@
   </script>
   <script type="module">
     // Load JS and attach instance to window.
+    import { teleport } from "@vrembem/core/plugins";
     window["vb"] = await import("./index.js");
 
     // Get all the theme set button elements.
@@ -56,7 +57,7 @@
     
     const modal = new vb.Modal({
       plugins: [
-        vb.core.teleport({
+        teleport({
           where: ".modals"
         })
       ]
@@ -64,7 +65,7 @@
 
     const drawer = new vb.Drawer({
       plugins: [
-        vb.core.teleport({
+        teleport({
           where: ".drawer-main",
           how: "before"
         })
